linux-qoriq-sdk.bb: prepend to do_configure
Instead of prepending to kernel_do_configure
lets prepend to do_configure this gives another
knob for users who want to insert between this
prepend and do_configure_kernel. Typical usecase
is to enable kconfig options on top of defconfig
can be then appended to .config easliy in a
bbappend.

Signed-off-by: Khem Raj <raj.khem@gmail.com>